pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "torchtrade"
version = "0.0.1"
description = "Reinforcement learning environments for trading applications"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Sebastian Dittert", email = "torchtradecontact@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "alpaca-py",
    "python-binance",
    "ccxt>=4.0.0", # Replaced python-bitget with ccxt for V2 API support
    "python-dotenv",
    "torchrl",
    "wandb",
    "tqdm",
    "hydra-core",
    "matplotlib",
    "ta",
    "datasets",
    "scikit-learn>=1.6.1",  # Used by example training scripts for StandardScaler
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
docs = [
    "mkdocs-material",
]
llm_local = [
    "vllm>=0.6.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.41.0",
]
llm = [
    "openai>=1.0.0",
]
chronos = [
    "chronos-forecasting",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[project.urls]
"Homepage" = "https://github.com/TorchTrade/torchtrade"
"Bug Reports" = "https://github.com/TorchTrade/torchtrade/issues"
"Source" = "https://github.com/TorchTrade/torchtrade"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["torchtrade"]
